todo_app/requirements.txt
Flask==1.1.1
python-dotenv==0.12.0
requests==2.24.0
pytest==6.0.1
python-dateutil==2.8.1
selenium==3.141.0